Refactor passes
