+++ lib/hir-mir/src/lib.rs
@@ -29,7 +29,7 @@ pub use compile_configuration::CompileConfiguration;
 pub use concurrency_configuration::ConcurrencyConfiguration;
 pub use error::CompileError;
 pub use error_type_configuration::ErrorTypeConfiguration;
-use hir::{analysis::types::type_existence_validator, ir::*};
+use hir::{analysis::type_existence_validator, ir::*};
 pub use list_type_configuration::ListTypeConfiguration;
 pub use main_module_configuration::*;
 pub use string_type_configuration::StringTypeConfiguration;
