+++ lib/hir-mir/src/type_compiler.rs
@@ -1,7 +1,7 @@
 use super::{context::CompileContext, CompileError};
 use fnv::FnvHashMap;
 use hir::{
-    analysis::types::{type_canonicalizer, type_id_calculator},
+    analysis::{type_canonicalizer, type_id_calculator},
     types::{self, Type},
 };
