+++ lib/hir-mir/src/type_inferrer.rs
@@ -1,7 +1,7 @@
 use super::{context::CompileContext, environment_creator, type_extractor, CompileError};
 use fnv::FnvHashMap;
 use hir::{
-    analysis::types::{type_canonicalizer, type_difference_calculator, union_type_creator},
+    analysis::{type_canonicalizer, type_difference_calculator, union_type_creator},
     ir::*,
     types::{self, Type},
 };
