+++ lib/hir/src/analysis.rs
@@ -1,2 +1,22 @@
-pub mod ir;
-pub mod types;
+pub mod definition_qualifier;
+mod error;
+pub mod expression_visitor;
+pub mod record_field_resolver;
+pub mod type_canonicalizer;
+pub mod type_collector;
+pub mod type_comparability_checker;
+pub mod type_difference_calculator;
+pub mod type_equality_checker;
+pub mod type_existence_validator;
+pub mod type_formatter;
+pub mod type_id_calculator;
+pub mod type_qualifier;
+pub mod type_resolver;
+pub mod type_subsumption_checker;
+pub mod type_transformer;
+pub mod union_type_creator;
+pub mod union_type_member_calculator;
+pub mod variable_renamer;
+pub mod variable_transformer;
+
+pub use error::TypeError;
