+++ lib/hir/src/analysis/error.rs
