--- lib/hir/src/analysis/ir.rs
@@ -1,6 +1,0 @@
-pub mod definition_qualifier;
-pub mod expression_visitor;
-pub mod type_qualifier;
-pub mod type_transformer;
-pub mod variable_renamer;
-pub mod variable_transformer;
