+++ lib/hir/src/analysis/type_collector.rs
