+++ lib/hir/src/analysis/type_formatter.rs
