+++ lib/hir/src/analysis/union_type_creator.rs
