+++ lib/hir/src/analysis/variable_renamer.rs
