+++ lib/hir/src/analysis/variable_transformer.rs
